abstract the join-specific functions into a separate base class
